+++ exerc06/seq-hists/seq-len.sh
@@ -1,0 +1,1 @@
+ecoz2 seq show $1 | sed -E 's/.*L=(.+)\).*/\1/g'